Game Summary UI
The game summary UI is now a thing. This is how you now generate a level, it no longer just makes one when you win/lose.

I still need to implement the time taken stat, then it will be added properly to the window.

A lot of bug fixing, some refactoring.